change zone value from int to float

+++ src/fontra/core/classes.py
@@ -93,7 +93,7 @@ class FontSource:
 @dataclass(kw_only=True)
 class FontMetric:
     value: float
-    zone: int = 0
+    zone: float = 0
     customData: CustomData = field(default_factory=dict)
